pyproject.toml
[tool.poetry]
name = "product-store"
version = "0.1.0"
description = ""
authors = ["jesustr20 <jesus.tr61@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.6"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.7.0"
openai = "^1.64.0"
langchain = "^0.3.19"
langchain-community = "^0.3.18"
langchain-openai = "^0.3.7"
langchain-core = "^0.3.39"
langsmith = "^0.3.10"
langgraph = "^0.2.74"
psycopg2-binary = "^2.9.10"
djangorestframework-simplejwt = "^5.4.0"
chromadb = "^0.6.3"
redis = "^5.2.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
